requirements/dev.txt
-r common.txt
autopep8
django-nose
django-debug-toolbar
django_extensions==1.3.8
Werkzeug==0.9.6
pytest-config==0.0.8
vcrpy==1.0.2
mock==1.0.1
model_mommy==1.2
